+++ BOJ/14003 가장 긴 증가하는 부분 수열 5.cpp
@@ -32,7 +32,7 @@ struct p14003
 		cin.tie(0);
 		cout.tie(0);
 
-		const int MAX = 1e6;
+		const int MAX = (int)1e6;
 		int a[MAX];
 		int d[MAX];
 		int s[MAX];
